Cleanup api models
