+++ src/api/models/datastores.ts
@@ -1,0 +1,8 @@
+export interface GetDatastoresElemResponse {
+  datastore_id: string;
+  name: string;
+  display_id: string;
+  deleted: boolean;
+  imported: boolean;
+  uploading: boolean;
+}
