+++ src/api/models/templates.ts
@@ -1,0 +1,15 @@
+export interface GetTemplatesTemplateResponse{
+  tp_id: string;
+  name: string;
+  description: string;
+}
+
+export interface GetTemplatesCategoryElemResponse{
+  category: string;
+  templates: GetTemplatesTemplateResponse[];
+}
+
+export interface GetTemplatesCategoryResponse{
+  enabled: boolean;
+  categories: GetTemplatesCategoryElemResponse[];
+}
