+++ src/commands/workspaces/get.ts
@@ -2,7 +2,7 @@ import {flags} from '@oclif/command'
 import {cli} from 'cli-ux'
 import chalk from 'chalk'
 import BaseWithContext from '../../base-with-context'
-import {GetWorkspacesResponse} from '../../api/workspaces/workspaces'
+import {GetWorkspacesResponse} from '../../api/models/workspaces'
 
 export default class WorkspacesGet extends BaseWithContext {
   static description = 'get workspaces from hexabase'
